Oracle/requirements.txt
firebase-admin
google-cloud-aiplatform
google-api-python-client
requests
Pillow